Refactor copy command
to separate code for creating the task when using the template
